fix: move tool-result media normalization into agent
