+++ apps/server/src/agent/compaction/content.ts
