+++ apps/server/src/agent/compaction/prompt.ts
@@ -5,7 +5,7 @@ import type {
   ToolResultPart,
   UserContent,
 } from 'ai'
-import { toolResultOutputToText } from './compaction-content'
+import { toolResultOutputToText } from './content'
 
 const SUMMARIZATION_SYSTEM_PROMPT = `You are a context summarization assistant. Your task is to read a conversation between a user and an AI assistant, then produce a structured summary following the exact format specified.
